backend/requirements.txt
# FastAPI Ecosystem
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7  # Required for pymysql with MySQL 8+
alembic==1.12.1

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Environment variables
python-dotenv==1.0.0

# Reporting
reportlab==4.0.7
openpyxl==3.1.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Desktop Application
pywebview==4.4.1

# HTTP Client
requests>=2.31.0
